+++ DNN_Project/CSV.h
@@ -6,7 +6,7 @@ using namespace std;
 static class CSV
 {
 public:
-	static list<list<float>> read(string file);
+	static list<list<double>> read(string file);
 private:
 
 };
